% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{collect_relative_risks}
\alias{collect_relative_risks}
\title{Collect files on relative risks in compact lists}
\usage{
collect_relative_risks(in_list, keyword_map)
}
\arguments{
\item{in_list}{A list of lists, where the upper most level are names of
risk factors or diseases, and each list contains one or multiple lists corresponding
to the names in \code{mapping}.}

\item{keyword_map}{A named list that maps list names on the second nesting of \code{in_list} (the keys) to
shorter names, or NULL (the values). If the values are not NULL, it is assigned to the \code{to}
value of the entire list in the output. If values are NULL, the names of the list
that are being iterated over are assigned to \code{to} in the output.}
}
\value{
A list with 2 elements, with the names coming from the names of \code{keyword_map}.
The two lists contain a list of 3, named \code{from}, \code{to}, and \code{filename}. They indicate
the name of the file that contains the relative risks for going from risk factors
to diseases, between diseases, or from diseases to death/disability. If for any
type of relative risk there is no file present, the list is empty.
}
\description{
The function iterates over the keys in \code{keyword_map}. For each key, it iterates
over the elements in \code{in_list}, extracts the element named by the key and
binds the dataframes into one, using rbind.
It requires that all keys in \code{keyword_map} are present as list names the lists of \code{in_list}.
}
\keyword{internal}
