% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{configure_disease}
\alias{configure_disease}
\title{Configure a disease}
\usage{
configure_disease(
  name,
  prevalence_filename,
  incidence_filename,
  excess_mortality_filename,
  disability_weights_filename
)
}
\arguments{
\item{name}{A character string which specifies the unique name of the disease.}

\item{prevalence_filename}{A character string which specifies the path to the prevalence file.}

\item{incidence_filename}{A character string which specifies the path to the incidence file.}

\item{excess_mortality_filename}{A character string which specifies the path to the
excess mortality file.}

\item{disability_weights_filename}{A character string which specifies the path to the
disability weights file.}
}
\value{
A list containing the disease configuration parameters.
}
\description{
Creates a configuration list for a disease in the simulation.
}
\examples{
configure_disease(
  name = "Lung_Cancer",
  prevalence_filename = "NL_disease_LungCa_Prev_V2",
  incidence_filename = "NL_disease_LungCa_Inc_V2",
  excess_mortality_filename = "NL_disease_LungCa_ExecMor_V2",
  disability_weights_filename = "NL_disease_LungCa_DALY_V1"
)

}
