% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{configure_risk_factor}
\alias{configure_risk_factor}
\title{Configure a risk factor}
\usage{
configure_risk_factor(name, transition_filename, prevalence_filename)
}
\arguments{
\item{name}{A character string which specifies the unique name of the risk factor.}

\item{transition_filename}{A character string which specifies the path to the
transition rates file.}

\item{prevalence_filename}{A character string which specifies the path to the prevalence file.}
}
\value{
A list containing the risk factor configuration parameters.
}
\description{
Creates a configuration list for a risk factor in the simulation.
}
\examples{
configure_risk_factor(
  name = "Smoking_cat3", # Should be the same as the name in the reference data
  transition_filename = "NL_RF_smoking_Transitions_Netto",
  prevalence_filename = "NL_RF_smoking_cat3_Prev_V1"
)

}
