% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_server.R
\name{create.shiny.server}
\alias{create.shiny.server}
\title{This creates the Shiny server.}
\usage{
create.shiny.server(
  risk.factor.prevalence,
  risk.factor.transition,
  relative.risks,
  relative.risk.other
)
}
\description{
This creates the Shiny server.
}
