% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{create.shiny.ui}
\alias{create.shiny.ui}
\title{This creates the Shiny UI.}
\usage{
create.shiny.ui(
  dynamodir,
  rundir,
  simdir,
  population,
  diseases,
  risk.factor,
  relative.risks
)
}
\value{
dashboardPage A Shiny dashboardPage() element.
}
\description{
This creates the Shiny UI.
}
