% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_choice_list}
\alias{create_choice_list}
\title{Create a choice list from a data frame of relative risks}
\usage{
create_choice_list(input_df, caller = "ui")
}
\arguments{
\item{input_df}{A data frame containing relative risk data.}

\item{caller}{A string indicating the calling context, either "ui" or "server".}
}
\value{
A named list where names are either full descriptions of the relative
risks (if \code{caller = "ui"}) or file names (if \code{caller = "server"}).
}
\description{
Create a choice list from a data frame of relative risks
}
\details{
This function generates a choice list from a data frame of relative risks.
The format of the returned list depends on whether it's being called
from the UI or server context.
}
\keyword{internal}
