% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_choices_from_relative_risks}
\alias{create_choices_from_relative_risks}
\title{Create choices from filtered relative risks}
\usage{
create_choices_from_relative_risks(relative_risks, caller = "ui")
}
\arguments{
\item{relative_risks}{A list containing available relative risk data.}

\item{caller}{A string indicating the calling context, either "ui" or "server".}
}
\value{
A list containing four elements:
\item{disease_choices}{A list of choices for relative risks from disease to disease}
\item{risk_factor_choices}{A list of choices for relative risks from risk factor to disease}
\item{death_choices}{A list of choices for relataive risks from risk factor to death}
\item{disability_choices}{A list of choices for relative risks from risk factor to disability}
}
\description{
Create choices from filtered relative risks
}
\details{
This function processes the available relative risk data and creates choice
lists. The lists can be used in UI elements such as checkBoxGroupInput.
If using \code{caller = "server"}, it can be used in the server listening to the
respective UI to return the file names of the elements chosen by the user.
}
\keyword{internal}
