% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_factors.R
\name{create_risk_factor_dir}
\alias{create_risk_factor_dir}
\title{Create directory with XML files for a risk factor}
\usage{
create_risk_factor_dir(
  risk_factor_name,
  transition_matrix_list,
  transition_drift_list,
  relative_risks_death_list,
  prevalences_list,
  prevalences_duration_list,
  relative_risks_disability_list,
  risk_factor_configuration
)
}
\arguments{
\item{risk_factor_name}{A character string specifying the name of the risk factor.
A directory will be created with this name if it does not already exist.}

\item{transition_matrix_list}{A list of transition matrices, where each element contains
a list with \code{'data'} and \code{'type'} components.}

\item{transition_drift_list}{A list of transition drifts, where each element contains
a list with \code{'data'}, \code{'type'}, and \code{'trend'} components.}

\item{relative_risks_death_list}{A list of relative risks for death, where each element
contains a list with \code{'data'} and \code{'type'} components.}

\item{prevalences_list}{A list of prevalences, where each element contains a list with
\code{'data'}, \code{'type'}, and \code{'distribution'} components.}

\item{prevalences_duration_list}{A list of prevalence durations, where each element
contains a list with a `'data'`` component.}

\item{relative_risks_disability_list}{A list of relative risks for disability, where
each element contains a list with \code{'data'} and \code{'type'} components.}

\item{risk_factor_configuration}{A list of risk factor configurations containing
\code{'data'}, \code{'reference'}, and \code{'type'} components.}
}
\value{
A logical value: TRUE if the directory structure was successfully created.
}
\description{
This function creates a directory structure for a risk factor and populates it with
various XML files containing risk factor data.
}
\details{
The function creates a directory named after the disease and subdirectories for:
\itemize{
\item Transitions
\item Relative_Risks_For_Disability
\item Relative_Risks_For_Death
\item Prevalences
\item DurationDistributions
\item configuration.xml
}

For each category, the function generates XML files based on the data provided in the corresponding input list.
Each XML file is saved in the appropriate subdirectory.

The transition matrix and drift type can be \code{'zero'} or \code{'netto'}.
The risk factor \code{'type'} can be \code{'continuous'}, \code{'categorical'}, or \code{'compound'}.
For \code{prevalences_list}, only \code{'continuous'} and \code{'categorical'} are valid types.
}
\examples{
\dontrun{
# A list element corresponds to a population
transition_matrix_list <- list('NL' = list(data = data.frame(...), type = 'zero'))
transition_drift_list <- list('NL' = list(data = data.frame(...), type = 'zero'))

prevalences_list <- list('NL' = list(data = data.frame(...), type = 'continuous', distribution = 'normal')
prevalences_duration_list <-  list('NL' = list(data = data.frame(...)))

# A list element corresponds to a risk factor
relative_risks_death_list <- list('continuous' = list(data = data.frame(...), type = 'continuous')
relative_risks_disability_list <-  list('continuous' = list(data = data.frame(...), type = 'continuous')

# Configuration for which risk factor element to use
list(data = data.frame(...), reference = list(referencevalue = 0.5), type = 'continuous')

create_risk_factor_dir(
  'smoking',
  transition_matrix_list = transition_matrix_list,
  transition_drift_list = transition_drift_list,
  relative_risks_death_list = relative_risks_death_list,
  prevalences_list = prevalences_list,
  prevalences_duration_list = prevalences_duration_list,
  relative_risks_disability_list = relative_risks_disability_list,
  risk_factor_configuration_list = risk_factor_configuration
)
}

}
