% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_github_release}
\alias{download_github_release}
\title{Download and extract GitHub release asset}
\usage{
download_github_release(
  repo_url = "https://github.com/Dynamo-HIA/dynamo-hia",
  release_tag = "v3.0.0-beta.1",
  os = NULL,
  dest_dir = tempdir(),
  make_executable = TRUE,
  github_pat = NULL
)
}
\arguments{
\item{repo_url}{Character string of the repository URL.
Default is \code{"https://github.com/Dynamo-HIA/dynamo-hia"}.}

\item{release_tag}{Character string of the release tag. Default is \code{"v3.0.0-beta.1"}.}

\item{os}{Optional character string of the operating system. Default is \code{NULL} which tries
to detect the OS automatically.}

\item{dest_dir}{Character string of the destination directory for extraction
(default: \code{tempdir()}).}

\item{make_executable}{Logical indicating whether to set execute permissions. Default is \code{TRUE}.}

\item{github_pat}{Optional GitHub Access Token. Default is \code{NULL} which will try to use the
token from the environment variable \code{GITHUB_PAT}.}
}
\value{
An invisible \code{fs::path} object pointing to the extracted contents.
}
\description{
Download and extract GitHub release asset
}
\examples{
\dontrun{
# With default arguments
download_github_release()

# From specific repository and release tag
download_github_release(
  repo_url = "https://github.com/Dynamo-HIA/dynamo-hia",
  release_tag = "v3.0.0-beta.1"
)
}

}
