% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{extract_relative_risk_source}
\alias{extract_relative_risk_source}
\title{Extract the source of relative risk transitions}
\usage{
extract_relative_risk_source(file_name)
}
\arguments{
\item{file_name}{The file name as a string.}
}
\value{
A string. An error is thrown if the file name is not in the expected format or does
not have the right ending.
}
\description{
Extracts the substring after the unique hyphen in the file name.
}
\examples{
\dontrun{
extract_relative_risk_source("RR_to_BreatCa-Smoking_dur.xml")
}
}
