% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_server_data}
\alias{fetch_server_data}
\title{Get return data from a list of servers}
\usage{
fetch_server_data(server_name_prefix, server_list, item_names)
}
\arguments{
\item{server_name_prefix}{The prefix for the server names. The function looks
for server names with \code{server_name_prefix} + \code{i}, where \code{i} ranges from 1 to
the number of servers in \code{server_list}.}

\item{server_list}{A list of shiny \code{moduleServer}s.}

\item{item_names}{The names of items which each server refers to. For instance,
these can be the names of diseases or risk factors.}
}
\value{
A named list of non-NULL server values.
}
\description{
Get return data from a list of servers
}
\keyword{internal}
