% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_df_from_server_data}
\alias{filter_df_from_server_data}
\title{Filter a dataframe based on return values from a list of servers.}
\usage{
filter_df_from_server_data(df_to_filter, server_list, filter_col = "filename")
}
\arguments{
\item{df_to_filter}{A data frame to be filtered by the output of \code{server_list}.}

\item{server_list}{A list of servers, each returning a character vector, for instance
of user-chosen elements.}

\item{filter_col}{The column in \code{df_to_filter} to use for subsetting relative to
user input data.}
}
\value{
A \code{data.frame}
}
\description{
Filter a dataframe based on return values from a list of servers.
}
\keyword{internal}
