% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_items}
\alias{filter_items}
\title{Filter a list of items}
\usage{
filter_items(
  input_item,
  selected_risk_factors,
  risk_factor_feature,
  items_to_filter,
  debug
)
}
\arguments{
\item{input_item}{A string with a chosen file name.}

\item{selected_risk_factors}{A list of two character vectors, named "transitions"
and "prevalence". The list is obtained from evaluating the output of
\link{risk_factor_server}.}

\item{risk_factor_feature}{A string of either "transitions" or "prevalence".}

\item{items_to_filter}{A list of lists to be filtered.}

\item{debug}{A boolean indicating debug mode. If TRUE, the function prints
some information messages to the console.}
}
\value{
A list of two: \code{keep_items} holds the reduced \code{items_to_filter}, and
\code{default_value} holds the value that can be passsed to the \code{selected} argument
in \link[shiny]{selectInput}. The \code{default_value} are defined
such that they do not cause an infinite loop if two \code{selectInput} fields
reactively depend on each other.
}
\description{
For a list of lists, returns for each element in the list the \code{setdiff} to
some input.
}
\details{
The function checks if the \code{input_item} is in the \code{selected_risk_factors} of
the respective \code{risk_factor_feature} (prevalences or transitions). If so,
it drops the selected risk factor files of \emph{the other} risk factor feature
from the list.
}
\keyword{internal}
