% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_relative_risks}
\alias{filter_relative_risks}
\title{Filter relative risks based on selected diseases and risk factors}
\usage{
filter_relative_risks(relative_risks, diseases, risk_factors)
}
\arguments{
\item{relative_risks}{A list containing relative risk data.}

\item{diseases}{A vector of selected disease names.}

\item{risk_factors}{A vector of selected risk factor names.}
}
\value{
A list containing data frames:
\item{from_diseases}{Relative risks between diseases}
\item{from_risk_factors}{Relative risks from risk factors to diseases}
\item{to_death}{Relative risks from risk factors to death}
\item{to_disability}{Relative risks from risk factors to disability}
}
\description{
Filter relative risks based on selected diseases and risk factors
}
\details{
This function filters the relative risk data based on the provided diseases
and risk factors. It returns a subset of the original data that only includes
the specified diseases and risk factors. If there are no relative risk for a
given category, it returns an empty data frame in that category.
}
\keyword{internal}
