% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{get_disease_filenames}
\alias{get_disease_filenames}
\title{Get disease filenames}
\usage{
get_disease_filenames(root_dir, disease_name)
}
\arguments{
\item{root_dir}{A character string which specifies the root directory of the project.}

\item{disease_name}{A character string which specifies the name of the disease directory.}
}
\value{
A character vector of filenames associated with the specified disease.
}
\description{
Retrieves all filenames associated with a specific disease.
}
\examples{
root_dir <- system.file("extdata", "example-nl", "Reference_Data", package = "hiaR")
get_disease_filenames(root_dir, "Lung_Cancer")

}
