% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{get_file_tree}
\alias{get_file_tree}
\title{Create a tree of nested files and directories}
\usage{
get_file_tree(tree_root)
}
\arguments{
\item{tree_root}{Path to the root of the tree.}
}
\value{
A nested list. The leaf nodes are either empty directories or (xml) file names.
}
\description{
Recurses through \code{tree_root} and returns a nested, named list
corresponding to the nesting structure of files. At the leaf nodes,
which can be empty directories or files, it returns NULL.
}
\keyword{internal}
