% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{get_population_filenames}
\alias{get_population_filenames}
\title{Get population filenames}
\usage{
get_population_filenames(root_dir, population_name)
}
\arguments{
\item{root_dir}{A character string which specifies the root directory of the project.}

\item{population_name}{A character string which specifies the name of the population directory.}
}
\value{
A character vector of filenames associated with the specified population.
}
\description{
Retrieves all filenames associated with a specific population.
}
\examples{
root_dir <- system.file("extdata", "example-nl", "Reference_Data", package = "hiaR")
get_population_filenames(root_dir, "Netherlands")

}
