% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{get_population_names}
\alias{get_population_names}
\title{Get population directory names}
\usage{
get_population_names(root_dir, return_paths = FALSE)
}
\arguments{
\item{root_dir}{A character string which specifies the root directory of the project.}

\item{return_paths}{A logical which determines whether to return full paths (TRUE) or
only directory names (FALSE).}
}
\value{
An invisible character vector of population directory names or paths.
}
\description{
Retrieves the names or paths of population directories within the project structure.
}
\examples{
root_dir <- system.file("extdata", "example-nl", "Reference_Data", package = "hiaR")
get_population_names(root_dir)
get_population_names(root_dir, return_paths = TRUE)

}
