% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{get_reference_data}
\alias{get_reference_data}
\title{Get all reference data}
\usage{
get_reference_data(root_dir)
}
\arguments{
\item{root_dir}{A string specifying the path to the root directory containing the files to be grouped.}
}
\value{
A list with four main components:
\itemize{
\item diseases: A nested list representing the file structure under the "Diseases" directory
\item populations: A nested list representing the file structure under the "Populations" directory
\item risk_factors: A nested list representing the file structure under the "Risk_Factors" directory
\item relative_risks: A list with two components:
\itemize{
\item diseases: Relative risks extracted from the diseases group
\item risk_factors: Relative risks extracted from the risk factors group
}
}
}
\description{
Function that loads all metadata of reference data for a model.
The output of the function can be used by the Shiny app to display
options to the user.
The function organizes files from a specified root directory into predefined groups:
diseases, populations, risk factors, and relative risks. It creates a nested list
structure representing the file hierarchy within each group.
}
\details{
The function uses predefined keywords to identify and extract relative risk information
from the diseases and risk factors groups. For diseases, it looks for "Relative_Risks_From_Risk_Factor"
and "Relative_Risks_From_Diseases". For risk factors, it searches for "Relative_Risks_For_Death"
and "Relative_Risks_For_Disability".
}
\examples{
\dontrun{
root_dir <- "../data/Tutorial_Data/Reference_Data"
file_groups <- get_reference_data(root_dir)
}
}
\seealso{
\code{\link{get_file_tree}}, \code{\link{get_relative_risk_source}}
}
