% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{get_risk_factor_filenames}
\alias{get_risk_factor_filenames}
\title{Get risk factor filenames}
\usage{
get_risk_factor_filenames(root_dir, risk_factor_name)
}
\arguments{
\item{root_dir}{A character string which specifies the root directory of the project.}

\item{risk_factor_name}{A character string which specifies the name of the risk factor directory.}
}
\value{
A character vector of filenames associated with the specified risk factor.
}
\description{
Retrieves all filenames associated with a specific risk factor.
}
\examples{
root_dir <- system.file("extdata", "example-nl", "Reference_Data", package = "hiaR")
get_risk_factor_filenames(root_dir, "Smoking_cat3")

}
