% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{getDiseaseData}
\alias{getDiseaseData}
\title{reads the filenames in the  disease directory and checks if they are OK (does not consider the relative risks)}
\usage{
getDiseaseData(diseaseDir)
}
\arguments{
\item{diseaseDir}{Full path to the directory of the disease to be checked}
}
\value{
a list of lists:  list(incList, prevList, emList, dwList) containing respectively
incidence files, prevalence files, excess mortality files and Dalyweigth files
attributes are valid (true/false), errors, warnings
#'
}
\description{
reads the filenames in the  disease directory and checks if they are OK (does not consider the relative risks)
}
\examples{
diseaseDir <- paste0(dirtestdata, '/Reference_Data/Diseases/',subdir)
}
