% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{getDiseaseList}
\alias{getDiseaseList}
\title{file that retrieves diseases from the reference data, split into valid and invalid disease
Invalid diseases are those where something is wrong with the underlying files
This list does not includes relative risks / oddsratios for death and disability,
those are read separately}
\usage{
getDiseaseList(dynamodir)
}
\arguments{
\item{dynamodir}{Dynamo-Hia work directory (full path)}
}
\value{
list with two elements:  list valid diseases, list of invalid disease
invalid list contains values on errors
}
\description{
file that retrieves diseases from the reference data, split into valid and invalid disease
Invalid diseases are those where something is wrong with the underlying files
This list does not includes relative risks / oddsratios for death and disability,
those are read separately
}
\examples{


  dynamodir <- dirtestdata
}
