% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{getMultipleFiles}
\alias{getMultipleFiles}
\title{Title}
\usage{
getMultipleFiles(basedir, subdir, xsdname)
}
\arguments{
\item{basedir}{full path to a base directory}

\item{subdir}{path relative to the base directory of the directory to be scraped}

\item{xsd}{name of the schema}
}
\value{
list containing 1. a list of valid files in the directory
2. a list of invalid files in the directory (with errors added as attributes)
}
\description{
Title
}
