% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{getPopulations}
\alias{getPopulations}
\title{Gets a list of valid and invalid populations}
\usage{
getPopulations(dynamodir)
}
\arguments{
\item{dynamodir}{full path to dynamo-HIA workdirectory}
}
\value{
list with two elements:  list valid populations, list of invalid populationns
}
\description{
Gets a list of valid and invalid populations
}
