% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{getRiskFactorData}
\alias{getRiskFactorData}
\title{Checks if the riskfactor directory seems OK (does not consider the relative risks)}
\usage{
getRiskFactorData(rfDir)
}
\arguments{
\item{rfDir}{Full path to the directory for the risk factor to be checked}
}
\value{
a list with 4 elemens:   prevalenceFiles , transitionFiles,nClasses , type
containing respectively
a list of prevalence files, a list of transition file names, number of classes of
the riskfactor and type of the riskcator
attributes are valid (true/false), errors, warnings
}
\description{
Checks if the riskfactor directory seems OK (does not consider the relative risks)
}
