% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{getRiskFactorList}
\alias{getRiskFactorList}
\title{file that retrieves riskfactors from the reference data, split into valid and invalid risk factors
this does not includes relative risks / oddsratios for death and disability}
\usage{
getRiskFactorList(dynamodir)
}
\arguments{
\item{dynamodir}{full path to the Dynamo-HIA work directory}
}
\value{
list with two elements:  list valid risk factors, list of invalid risk factors
invalid list contains values on errors, both give the type of risk factor
}
\description{
file that retrieves riskfactors from the reference data, split into valid and invalid risk factors
this does not includes relative risks / oddsratios for death and disability
}
