% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{getRRTable}
\alias{getRRTable}
\title{Gets a dataframe containing the relative risk files for valid risk factors and diseases}
\usage{
getRRTable(dynamodir, disList, rfList)
}
\arguments{
\item{dynamodir}{workdirectory Dynamo-Hia (full path)}

\item{disList}{the list with diseases}

\item{rfList}{the list with riskfactors}
}
\value{
a dataframe with variables from (disease or riskfactor), to (disease, death or disability),
filename (excluding path), valid (TRUE/FALSE)
Only the relative risks pertaining to valid diseases and riskfactors are included.
}
\description{
Gets a dataframe containing the relative risk files for valid risk factors and diseases
}
