% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{getSingleFile}
\alias{getSingleFile}
\title{validate a single file with a known name}
\usage{
getSingleFile(currentpath, filename, xsdname)
}
\arguments{
\item{currentpath}{directory to read from}

\item{filename}{name to read (excluding .xml)}

\item{xsdname}{name of the schema to validate the file}
}
\value{
list containing 1. a list of the valid file in the directory
2. a list of the invalid file in the directory (with errors added as attributes)
although the list always only contain a single or no files, this structure is compatable with structures
for other parts where multiple files are possible
}
\description{
validate a single file with a known name
}
\examples{
currentpath <- "c:/temp/test/Reference_Data/Populations/testcountry"
filename <- "size"
xsdname <- "populationsize"
getSingleFile(currentpath, filename, xsdname)

currentpath <- "c:/temp/test/Reference_Data/Diseases/dis/Prevalences"
filename <- "testfile"
xsdname <- "diseaseprevalences"

}
