% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{getTransitionFiles}
\alias{getTransitionFiles}
\title{Reads and validates the riskfactor transitionfiles}
\usage{
getTransitionFiles(currentpath)
}
\arguments{
\item{currentpath}{full path to the directory with riskfactor data}
}
\value{
list containing 1. list van valid transitionfiles 2. list of invalid transition files
}
\description{
Reads and validates the riskfactor transitionfiles
}
\examples{

currentpath <-  paste0(rfDir)
}
