% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{getValidTree}
\alias{getValidTree}
\title{Get valid data in the Dynamo-Hia home directory}
\usage{
getValidTree(dynamodir)
}
\arguments{
\item{dynamodir}{full path to the Dynamo-HIA workdirectory (parent directory of the subdirectories
Reference_Data and Simulations)}
}
\value{
list with valid information in the tree. This list has 4 elements:
$populations containing a list of valid populations
$disease containing a list of valid diseases
$riskfactors containing a list of valid riskfactors
$relativeRisks containing a data.frame with relative risks
Populuations, diseases and riskfactors with incomplete data (particular files missing)
are invalid and not in this list
}
\description{
Get valid data in the Dynamo-Hia home directory
}
\examples{
getValidTree(dynamodir)
}
