% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{launch_application}
\alias{launch_application}
\title{Starts the R shiny app to define the settings and then run the HIA Java app.}
\usage{
launch_application()
}
\value{
A graphical user interface with the Shiny app.
}
\description{
Starts the R shiny app to define the settings and then run the HIA Java app.
}
\examples{
\dontrun{
launch_application()
}
}
