% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{load_simulation_results}
\alias{load_simulation_results}
\title{Load results from DYNAMO-HIA simulation}
\usage{
load_simulation_results(root_dir, simulation_name, type = "batchoutput")
}
\arguments{
\item{root_dir}{Character string specifying the path to the working directory.}

\item{simulation_name}{Character string specifying the name of the simulation.}

\item{type}{Character string specifying the type of results to load.
Default is \code{"batchoutput"} (population data). Other options are \code{"cohortLE"}
(life expectancy data) and \code{"sullivan"} (Sullivan's method data).}
}
\value{
A data frame containing the simulation results.
}
\description{
Load results from DYNAMO-HIA simulation
}
