% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{makeTreeList}
\alias{makeTreeList}
\title{Makes a list of the valid information in the Dynamo-HIA work directory. It also contains information on invalid files in the population,
risk factor and disease directories. Relative risk files are only included for valid disease and valid risk factors.}
\usage{
makeTreeList(dynamodir)
}
\arguments{
\item{dynamodir}{name of Dynamo-HIA workdirectory (parent directory of the subdirectories Reference_Data and Simulations). This is the full path}
}
\value{
A nested list with on the lowest level the file names of xml files, divided in lists of valid and invalid files.
The list contains on the highest level 1) list of populations, 2) list of diseases 3) list of risk factors
4) dataframe containing names of relative risk files associated with valid diseases and valid risk factors
The attribute "valid" is TRUE when the is at least one valid population and one valid risk factor
}
\description{
Makes a list of the valid information in the Dynamo-HIA work directory. It also contains information on invalid files in the population,
risk factor and disease directories. Relative risk files are only included for valid disease and valid risk factors.
}
\examples{
dynamodir <-   dir(system.file("extdata", package = "hiaR"))
treeList <- makeTreeList(dynamodir)
attributes(treeList)$valid
# if not find errors
attributes(treeList, "errors")
# check if the population is valid
attr(treeList$populations, "errors")
}
