% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{makeXML.age.sex.cat}
\alias{makeXML.age.sex.cat}
\title{###############################################################################
#
XML writing functions                               #
#
###############################################################################
Make an DYNAMO-HIA XML file with data per category, where the value is a value}
\usage{
makeXML.age.sex.cat(tag1, tag2, data, ncat, varname)
}
\arguments{
\item{tag1}{first level tag of XML}

\item{tag2}{second level tag of XML}

\item{data}{dataframe with variables cat, age, sex and the value to be
used (in a variable with a name as given by varname)
dataframe with variables age, sex, value to be used (as given by varname)
where sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{ncat}{number of categories}

\item{varname}{the name of the variable with the value  (in quotes)}
}
\value{
string containing the xml
}
\description{
###############################################################################
#
XML writing functions                               #
#
###############################################################################
Make an DYNAMO-HIA XML file with data per category, where the value is a value
}
