% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{makeXML.age.sex.cat2}
\alias{makeXML.age.sex.cat2}
\title{Make an DYNAMO-HIA XML file with data per category, where the value is a percentage}
\usage{
makeXML.age.sex.cat2(tag1, tag2, data, ncat, varname)
}
\arguments{
\item{tag1}{first level tag of XML}

\item{tag2}{second level tag of XML}

\item{data}{dataframe with variables cat (categories of riskfactor) age, sex, value to be used (as given by varname)
where cat is coded with values 1 to ncat (without gaps), sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{ncat}{number of categories}

\item{varnamevarname}{the name of the variable with the value (in quotes) to be written}
}
\value{
string containing the xml
}
\description{
Make an DYNAMO-HIA XML file with data per category, where the value is a percentage
}
