% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{makeXML.age.sex.compound}
\alias{makeXML.age.sex.compound}
\title{####################  for compound relative risk (general)       #####################
Writes the DYNAMO_HIA input XML file containing the relative risk of a categorical risk factor on diseae}
\usage{
makeXML.age.sex.compound(
  tag1,
  tag2,
  ncat,
  durcat,
  data,
  varname,
  varend,
  alpha
)
}
\arguments{
\item{tag1}{main tag of xml}

\item{tag2}{main tag of xml}

\item{ncat}{number of categories in the riskfactor}

\item{durcat}{number of the category with duration dependent relative risk}

\item{data}{dataframe with variables cat (categories of riskfactor), age, sex, value to be used (as given by varname)
where cat is coded 1 to ncat (consecutive) sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{name of variable that contains the value of the relative risk
for the duration category this is the beginvalue #' relative risk on the disease for each category of the risk factor}

\item{varend, }{name of variable that contains the value of the RR at infinity}

\item{alpha, }{name of variable that contains the value for alpha}
}
\value{
string with xml content
}
\description{
####################  for compound relative risk (general)       #####################
Writes the DYNAMO_HIA input XML file containing the relative risk of a categorical risk factor on diseae
}
\examples{
todo
}
