% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{makeXML.age.sex.duration}
\alias{makeXML.age.sex.duration}
\title{Make an DYNAMO-HIA XML file with data per duration, where the value is a value}
\usage{
makeXML.age.sex.duration(data, varname)
}
\arguments{
\item{data}{dataframe with variables duration, age, sex and the value to be
used (in a variable with a name as given by varname)
dataframe with variables age, sex, value to be used (as given by varname)
where sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname}{the name of the variable with the value  (in quotes)}

\item{tag1}{first level tag of XML}

\item{tag2}{second level tag of XML}
}
\value{
string containing the xml
}
\description{
Make an DYNAMO-HIA XML file with data per duration, where the value is a value
}
