% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{makeXML.age.sex3}
\alias{makeXML.age.sex3}
\title{Make an DYNAMO-HIA input ML file with a number (integer) as the content}
\usage{
makeXML.age.sex3(tag1, tag2, data, varname)

makeXML.age.sex3(tag1, tag2, data, varname)
}
\arguments{
\item{tag1}{first level tag of XML}

\item{tag2}{second level tag of XML}

\item{data}{dataframe with variables age, sex, value to be used (as given by varname)
where sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{name of the variable (in quotes) to be written}
}
\value{
string containing the xml

string containing the xml
}
\description{
Make an DYNAMO-HIA input ML file with a number (integer) as the content

Make an DYNAMO-HIA input ML file with a number (integer) as the content
}
