% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{makeXML.rf.configuration}
\alias{makeXML.rf.configuration}
\title{produces a string to be written to as configuration file for a categorical or compound risk factor}
\usage{
makeXML.rf.configuration(
  nclass,
  classnames,
  compound,
  referenceclass = 1,
  durationclass
)
}
\arguments{
\item{nclass}{number of categories in the risk factor}

\item{classnames}{names of the categories}

\item{compound}{boolean: whether compound or not}

\item{referenceclass}{reference category (category for which the relative risk is equal to 1)}

\item{durationclass}{for compound risk factors: the number of the category for which the relative risk
depends on how long the persons has been in the particular risk category,}
}
\value{
string to be written as configuration file for a categorical/compound risk factor (xml)
}
\description{
produces a string to be written to as configuration file for a categorical or compound risk factor
}
