% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{makeXML.rf.configuration2}
\alias{makeXML.rf.configuration2}
\title{produces a string to be written to as configuration file for a continuous risk factor}
\usage{
makeXML.rf.configuration2(cutoffs, referencevalue = 0)
}
\arguments{
\item{cutoffs}{Cut-off values used in output}

\item{referencevalue}{reference value (value of the risk factor for which the relative risk is equal to 1)}
}
\value{
string to be written as configuration file for a continuous risk factor (xml)
}
\description{
produces a string to be written to as configuration file for a continuous risk factor
}
