% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{makeXML.simulation.configuration}
\alias{makeXML.simulation.configuration}
\title{writes a string with the simulation configuration file}
\usage{
makeXML.simulation.configuration(
  newborns,
  startingyear,
  yearsSimulated,
  popSize,
  minAge,
  maxAge,
  refScenarioName,
  randomSeed,
  populationName,
  scenarioList,
  diseaseList,
  riskfactorList,
  relriskList
)
}
\arguments{
\item{newborns}{0: no newborns added during simulation 1: newborns added}

\item{startingyear}{startyear of simulation}

\item{yearsSimulated}{number of years to simulate}

\item{popSize}{number of simulated persons per age and sex group}

\item{minAge}{minimum age of simualted population}

\item{maxAge}{maximum age of simulated population}

\item{refScenarioName}{name of the reference (business as usual) scenario}

\item{randomSeed}{random seed used in the simulation}

\item{populationName}{name of the population to be used}

\item{scenarioList}{list with one element per scenario. Each element has the following named objects:
name, successRate, minAge, maxAge,  targetSex, transitionFile, prevalenceFile}

\item{diseaseList}{list with one element per disease. Each element has the following named objects:
name, prevalenceFile, incidenceFile, excessMortFile, disabilityFile}

\item{riskfactorList}{list with named elements: name, transitionFile, prevalenceFile}

\item{relriskList}{list with one element per relative risk. Each element has the following named elements:
from, to, filename}
}
\description{
writes a string with the simulation configuration file
}
