% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_population_pyramid}
\alias{plot_population_pyramid}
\title{Create a Population Pyramid Visualization}
\usage{
plot_population_pyramid(
  df,
  year,
  scenario,
  difference = FALSE,
  split_gender = TRUE,
  show_total = TRUE,
  show_disability = TRUE,
  show_disease = TRUE,
  alpha = 0.8
)
}
\arguments{
\item{df}{A data frame containing batch output results from DYNAMO-HIA.}

\item{year}{The specific year to plot.}

\item{scenario}{The index of the scenario to plot.}

\item{difference}{Logical, whether to show differences from reference scenario (default\code{ FALSE}).}

\item{split_gender}{Logical, whether to split data by gender (default \code{TRUE}).}

\item{show_total}{Logical, whether to display total population (default \code{TRUE}).}

\item{show_disability}{Logical, whether to display disability data (default \code{TRUE}).}

\item{show_disease}{Logical, whether to display disease data (default \code{TRUE}).}

\item{alpha}{Transparency level for plot elements (default 0.8).}
}
\value{
A \code{ggplot} object representing the population pyramid.
}
\description{
Create a Population Pyramid Visualization
}
