% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_prevalence}
\alias{plot_prevalence}
\title{Plot Disease Prevalence Rates}
\usage{
plot_prevalence(
  df,
  disease,
  x_axis = "year",
  split_by = "scenario",
  age = NULL,
  year = NULL
)
}
\arguments{
\item{df}{A data frame containing batch output results from DYNAMO-HIA.}

\item{disease}{Disease to plot prevalence rates for.}

\item{x_axis}{Column to use for x-axis (default \code{"year"}).}

\item{split_by}{Column to use for color/fill splitting (default \code{"scenario"}).}

\item{age}{Optional age group to filter. By default (\code{NULL}), the youngest age group is used.}

\item{year}{Optional year to filter. By default (\code{NULL}), the earliest year is used.}
}
\value{
A ggplot object representing prevalence rates.
}
\description{
Plot Disease Prevalence Rates
}
