% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_risk_factor}
\alias{plot_risk_factor}
\title{Plot Risk Factor Prevalence Rates}
\usage{
plot_risk_factor(
  df,
  risk_factor_class,
  x_axis = "year",
  split_by = "scenario",
  age = NULL,
  year = NULL
)
}
\arguments{
\item{df}{A data frame containing batch output results from DYNAMO-HIA.}

\item{risk_factor_class}{Class of the risk factor to plot.}

\item{x_axis}{Column to use for x-axis (default \code{"year"}).}

\item{split_by}{Column to use for color/fill splitting (default \code{"scenario"}).}

\item{age}{Optional age group to filter. By default (\code{NULL}), the youngest age group is used.}

\item{year}{Optional year to filter. By default (\code{NULL}), the earliest year is used.}
}
\value{
A ggplot object representing prevalence rates.
}
\description{
Plot Risk Factor Prevalence Rates
}
