% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_config_module.R
\name{program_config_server}
\alias{program_config_server}
\title{Server part for program config}
\usage{
program_config_server(id)
}
\arguments{
\item{id}{The module namespace id, matching the \code{id} in the UI part.}
}
\value{
A named list of reactive values with the following elements
\itemize{
\item reference_data: A list loaded with \link{get_reference_data}.
\item dynamo_path: Path to the DYNAMO executable.
\item working_path: Working directory for the DYNAMO program.
}
}
\description{
The function observes the values in the UI, loads reference meta data and
sends them back as reactive values to the caller.
}
\keyword{internal}
