% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_config_module.R
\name{program_config_ui}
\alias{program_config_ui}
\title{UI part for program config}
\usage{
program_config_ui(id)
}
\arguments{
\item{id}{The id for the namespace of the module.}
}
\value{
Shows an error message when the reference data is not NULL and is
does not have the right content.
}
\description{
Creates two panels for the user to choose the paths to the DYNAMO executable
and to the working directory.
}
\keyword{internal}
