% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_module.R
\name{relative_risk_server}
\alias{relative_risk_server}
\title{Server Logic for Relative Risk Selection}
\usage{
relative_risk_server(id, relative_risks)
}
\arguments{
\item{id}{A character string. The ID that matches the corresponding UI element.}

\item{relative_risks}{A reactive expression returning the relative risks data.}
}
\value{
A reactive expression that returns a list
containing the names of the xml files corresponding to the selected
diseases and risk factors.
}
\description{
Server-side logic for selecting relative risks for diseases.
}
\keyword{internal}
