% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_factor_module.R
\name{risk_factor_server}
\alias{risk_factor_server}
\title{Server logic for managing multiple risk factors}
\usage{
risk_factor_server(id, reference_data)
}
\arguments{
\item{id}{A character string. The ID used to namespace the server logic, matching the ID used in the UI.}

\item{reference_data}{A reactive expression that returns a list containing risk factor information.
The list should have a 'risk_factors' element, which is itself a list where each element
represents a risk factor and contains file options for prevalences and transitions.}
}
\value{
A reactive expression that returns a named list of active risk factors and their selected values.
Each element in the list corresponds to an active risk factor and contains a nested list with
`prevalence' and 'transitions'.
}
\description{
Dynamically create and manage individual risk factor servers based on the provided reference data,
and collect user inputs for all active risk factors.
}
\keyword{internal}
