% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_factor_module.R
\name{risk_factor_ui}
\alias{risk_factor_ui}
\title{Create UI for Multiple Risk Factors}
\usage{
risk_factor_ui(id, reference_data)
}
\arguments{
\item{id}{A character string. The ID used to namespace the UI elements.}

\item{reference_data}{A reactive expression that returns a list containing risk factor information.
The list should have a 'risk_factors' element, which is itself a list where each element
represents a risk factor and contains file options for prevalences and transitions.}
}
\value{
A tagList containing UI components for each risk factor.
}
\description{
Create a collection of individual risk factor UI elements based on the
provided reference data.
}
\keyword{internal}
