% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dynamo_hia.R
\name{run_dynamo_hia}
\alias{run_dynamo_hia}
\title{Run the DYNAMO-HIA simulation}
\usage{
run_dynamo_hia(
  simulation_filename,
  app_filename,
  log_filename = "run_dynamo_hia.log"
)
}
\arguments{
\item{simulation_filename}{Path to the simulation batch file. Each line of this file must contain
a directory name of a simulation that should be run. This file should be located in the
"working directory", i.e., a directory that contains the \code{Reference_Data} and \code{Simulations}
subdirectories.}

\item{app_filename}{Path to the Java application.}

\item{log_filename}{Path to a log file. The default filename is \code{run_dynamo_hia.log}.
Can also be \code{FALSE} to disable logging.}
}
\value{
Returns \code{TRUE} if the simulations ran successfully. Otherwise, it will forward the
exception returned by the Java application.
}
\description{
A wrapper for running the Java application. Runs the application in batch mode.
}
