% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions_for_DYNAMO.R
\name{runDynamoHIA}
\alias{runDynamoHIA}
\title{Function running the java version of Dynamo-HIA}
\usage{
runDynamoHIA(dynamodir, simulationName, runDir = "extdata/application")
}
\arguments{
\item{dynamodir}{Full path of Dynaomo-HIA work directory (containing a valid Reference_Data and Simulations subdirectory)}

\item{simulationName}{Name of simulation (equal to a directory name  under "Simulations")#'}

\item{runDir}{name of the directory containing the Dynamo-HIA application#'}
}
\value{
results of run
}
\description{
Function running the java version of Dynamo-HIA
}
