% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_module.R
\name{scenario_ui}
\alias{scenario_ui}
\title{Create UI for scenarios}
\usage{
scenario_ui(id)
}
\arguments{
\item{id}{A character string. The ID used to namespace the UI element.}
}
\value{
A UI with 3 elements: an \code{actionButton} for adding and removing
scenarios, and a \code{tabsetPanel}. The panel can be populated with multiple
tabs, as handled by the server.
}
\description{
Create UI for scenarios
}
\keyword{internal}
