% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_config_module.R
\name{simulation_config_server}
\alias{simulation_config_server}
\title{Server part for simulation config}
\usage{
simulation_config_server(
  id,
  reference_data,
  disease_configs,
  risk_factor_configs,
  relative_risk_configs,
  scenario_configs,
  program_config
)
}
\arguments{
\item{id}{The module namespace id, matching the \code{id} in the UI part.}

\item{reference_data}{A reactive, containing the reference meta data.}

\item{disease_configs}{A reactive, containing the disease configurations.}

\item{risk_factor_configs}{A reactive, containing the risk factor configurations.}

\item{relative_risk_configs}{A reactive, containing the relative risk configurations.}

\item{scenario_configs}{A reactive, containing the scenario configurations.}

\item{program_config}{A reactive, containing the program configurations.}
}
\value{
A reactive list of key-value pairs read from the config UI, to be
used in the app/by other modules.
}
\description{
Processes all user input data from the UI into a reactive list.
Uses some reference meta data for letting the user choose between options.
When the \code{save} button is clicked, it will start and run the Java app
(not yet implemented).
}
\keyword{internal}
