% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_config_module.R
\name{simulation_config_ui}
\alias{simulation_config_ui}
\title{UI part for simulation config}
\usage{
simulation_config_ui(id)
}
\arguments{
\item{id}{The id for the namespace of the module.}

\item{population_choices}{A character vector with possible choices of
populations.}
}
\description{
UI part for simulation config
}
\keyword{internal}
