% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_selection_module.R
\name{single_disease_server}
\alias{single_disease_server}
\title{Server part for a single disease}
\usage{
single_disease_server(id)
}
\arguments{
\item{id}{A string. The module namespace id, matching the \code{id} in the UI part.}
}
\value{
A reactive expression. If the checkbox in the corresponding UI
is not checked, it returns \code{NULL}. Otherwise, it returns a list of 4:
\itemize{
\item prevalence
\item incidence
\item excess_mortality
\item disability
}
Each of the items contain the xml file the user chose in the UI.
}
\description{
Server part for a single disease
}
\keyword{internal}
