% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_selection_module.R
\name{single_disease_ui}
\alias{single_disease_ui}
\title{UI for selecting a single disease}
\usage{
single_disease_ui(id, disease_name, disease_files)
}
\arguments{
\item{id}{A string. The id for the namespace in the module}

\item{disease_name}{The name of the disease, displayed in the card title}

\item{disease_files}{A named list of xml files the user can choose from.
The list has 4 elements:
\itemize{
\item Prevalences
\item Incidences
\item Excess_Mortalities
\item Disability
}
Each element itself is a named list, where the names are xml files.}
}
\value{
A \code{bslib} card where the header is the disease name with a checkbox.
If the checkbox is clicked, the user sees the card content: \code{selectInput}
fields for prevalence, incidence, excess mortality and disability.
}
\description{
UI for selecting a single disease
}
\keyword{internal}
