% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_module.R
\name{single_relative_risk_ui}
\alias{single_relative_risk_ui}
\title{Create UI for a single group of relative risks.}
\usage{
single_relative_risk_ui(id, title, choices)
}
\arguments{
\item{id}{A character string. The ID used to namespace the UI elements.}

\item{title}{A character string. The title of the group of relative risks.}

\item{choices}{A reactive expression returning a list of possible choices for the checkbox group.}
}
\value{
A Shiny UI component (bslib::card) containing a checkbox group for the specified relative risk group.
}
\description{
Generate a UI component for a single relative risk group,
creating a card with a checkbox group.
}
\keyword{internal}
