% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_factor_module.R
\name{single_risk_factor_ui}
\alias{single_risk_factor_ui}
\title{Create UI for a Single Risk Factor}
\usage{
single_risk_factor_ui(id, risk_factor_name, risk_factor_files)
}
\arguments{
\item{id}{A character string. The ID used to namespace the UI elements.}

\item{risk_factor_name}{A character string. The name of the risk factor to be displayed.}

\item{risk_factor_files}{A list containing three named elements:
\itemize{
\item Prevalences: A named list of prevalence file options.
\item Transitions: A named list of transition file options.
}}
}
\value{
A Shiny UI component (bslib::card) for the specified risk factor.
}
\description{
Generate a UI component for a single risk factor in a Shiny application.
It creates a card with a checkbox header and conditional inputs for prevalence
and transitions.
}
\details{
The function creates a collapsible card with a checkbox in the header.
When the checkbox is checked, it reveals select inputs for prevalence and
transitions. The options for these inputs are populated
from the provided risk_factor_files list.
}
\seealso{
\code{\link{single_risk_factor_server}} for the corresponding server logic.
}
