% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_module.R
\name{single_scenario_ui}
\alias{single_scenario_ui}
\title{Create UI for single scenario}
\usage{
single_scenario_ui(id)
}
\arguments{
\item{id}{A character string. The ID used to namespace the UI elements.}
}
\value{
A Shiny UI component for defining a scenario.
}
\description{
Create UI for single scenario
}
\keyword{internal}
