% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_n_substrings}
\alias{validate_n_substrings}
\title{Validate that a string has N substrings of a given pattern}
\usage{
validate_n_substrings(s, pattern, count)
}
\arguments{
\item{s}{The string to check.}

\item{pattern}{The pattern to search for.}

\item{count}{The number of occurrences of \code{pattern} in \code{s}.}
}
\value{
Throws an error if the check fails.
}
\description{
Validate that a string has N substrings of a given pattern
}
\examples{
\dontrun{
validate_n_substrings("hello-world", "-", 1) # passes
validate_n_substrings("hello-world", "!", 1) # fails 
}
}
