% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_xml_schema}
\alias{validate_xml_schema}
\title{Validate an XML document against a schema}
\usage{
validate_xml_schema(root, schema_name)
}
\arguments{
\item{root}{An XML document object representing the root of the XML to be validated.}

\item{schema_name}{A character string representing the base name of the XML schema (with or without the \code{.xsd} extension) against which the XML document will be validated.}
}
\value{
This function returns nothing on success but throws an error if the XML document fails validation.
}
\description{
This function validates an XML document against a specified XML schema. If the XML document does not conform to the schema, the function throws an error with a message generated from the schema name.
}
\details{
The function reads the specified XML schema using \code{read_xml_schema} and validates the \code{root} XML document against it using \code{\link[xml2]{xml_validate}}.
}
\examples{
\dontrun{
# Validate an XML document against the 'example_schema' schema
root <- xml2::read_xml('<root><child>Example</child></root>')
validate_xml_schema(root, 'example_schema')
}

}
\keyword{internal}
