% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_tooltip}
\alias{wrap_tooltip}
\title{Wrap around a bslib tooltip}
\usage{
wrap_tooltip(element, description = "", placement = "auto")
}
\arguments{
\item{element}{An arbitrary, fully specified UI element.}

\item{description}{The description to show as a tooltip.}

\item{placement}{The position of the description. Passed as argument
with the same name to \code{tooltip}. Default is "auto".}
}
\value{
A \code{\link[bslib]{tooltip}} expression to be used in an UI element.
}
\description{
Specify a UI element together with a description.
}
\keyword{internal}
