% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{write_population_dir}
\alias{write_population_dir}
\title{Write XML files for a population to a directory}
\usage{
write_population_dir(
  population_name,
  sex_ratio,
  starting_year,
  newborns_df,
  size_df,
  mortality_df,
  disability_df
)
}
\arguments{
\item{population_name}{A character string representing the name of the population.
A directory will be created with this name if it does not already exist.}

\item{sex_ratio}{A numeric value representing the sex ratio of newborns (male to female ratio).}

\item{starting_year}{An integer representing the starting year of the newborns data.}

\item{newborns_df}{A data frame containing data for newborns, typically with columns for the
year and the number of newborns.}

\item{size_df}{A data frame containing population size data, typically with columns for the
year and population size.}

\item{mortality_df}{A data frame containing overall mortality data, typically with columns for
the year and mortality rate.}

\item{disability_df}{A data frame containing overall disability data, typically with columns for
the year and disability rate.}
}
\value{
Returns \code{TRUE} if the directory and XML files were created successfully.
}
\description{
This function creates a directory structure for a specified population and generates XML files
containing data for newborns, population size, overall mortality, and disability.
The files are saved in the population directory. The function only creates
files for provided arguments and leaves other files as they are.
}
\details{
The function creates a directory named after the population and generates the following
XML files within it:
\itemize{
\item{\code{newborns.xml}: Contains data for newborns, including the sex ratio and number of newborns
per year.}
\item{\code{size.xml}: Contains data for population size per year.}
\item{\code{overallmortality.xml}: Contains overall mortality rates per year.}
\item{\code{overalldisability.xml}: Contains overall disability rates per year.}
}

The function uses helper functions to create XML documents for each type of data.
These functions are expected to handle the formatting and structuring of the data into valid XML.
}
\examples{
\dontrun{
newborns <- data.frame(...)
size <- data.frame(...)
mortality <- data.frame(...)
disability <- data.frame(...)

create_population_dir("ExamplePopulation", 1.05, 2009, newborns, size, mortality, disability)
}

}
\seealso{
\code{\link{create_disease_dir}}
}
