% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeDisDaly}
\alias{writeDisDaly}
\title{####################   function to write     ###############################
####################  disease DALY-weight #####################################
####################  this delivers DYNAMO-1 data ###########################
Writes the DYNAMO_HIA input XML file containing the DALY weights for a disease in DANOMO-I format (obsolete)}
\usage{
writeDisDaly(diseasename, dynamodir, filename, data, varname)
}
\arguments{
\item{diseasename}{Disease name as used in DYNAMO-HIA}

\item{dynamodir}{= DYNAMO-HIA work directory}

\item{filename}{}

\item{data}{dataframe with variables age, sex, value to be used (as given by varname)
were sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname}{name of the variable in the dataframe that contains the values to be put
in the file, namely the DALY weight (values between 0 and 100)}
}
\value{
this is not exported to prevent users using the wrong function
}
\description{
####################   function to write     ###############################
####################  disease DALY-weight #####################################
####################  this delivers DYNAMO-1 data ###########################
Writes the DYNAMO_HIA input XML file containing the DALY weights for a disease in DANOMO-I format (obsolete)
}
