% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeDiseaseDisability}
\alias{writeDiseaseDisability}
\title{######################   function to write     ###############################
####################  disease disability  #####################################
this delivers  DYNAMO-2 data #############################################
Writes the DYNAMO_HIA input XML file containing the DALYweights / percentage disability for a disease}
\usage{
writeDiseaseDisability(diseasename, dynamodir, filename, data, varname)
}
\arguments{
\item{diseasename}{Disease name as used in DYNAMO-HIA}

\item{dynamodir}{DYNAMO-HIA work directory}

\item{filename}{name of the xml file that contains the DALYweights / percentage disability (without the .xml extension)}

\item{data}{dataframe with variables age, sex, value to be used (as given by varname)
were sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{namelijk the percentage disability caused bythe disease
in the file}
}
\description{
######################   function to write     ###############################
####################  disease disability  #####################################
this delivers  DYNAMO-2 data #############################################
Writes the DYNAMO_HIA input XML file containing the DALYweights / percentage disability for a disease
}
