% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeDisPrevalence}
\alias{writeDisPrevalence}
\title{#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
######################################################################
#####################   function to write disease prevalence   #######
######################################################################
Writes the DYNAMO_HIA input XML file containing the prevalence of a disease}
\usage{
writeDisPrevalence(diseasename, dynamodir, filename, data, varname)
}
\arguments{
\item{diseasename}{Disease name as used in DYNAMO-HIA}

\item{dynamodir}{= DYNAMO-HIA work directory}

\item{filename}{Filename under which to write the disease prevalence XML file (without XML extension)}

\item{data}{dataframe with variables age, sex, value to be used (as given by varname)
were sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{name of variable that contains the value to put in file, namely the disease prevalence}
}
\description{
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
######################################################################
#####################   function to write disease prevalence   #######
######################################################################
Writes the DYNAMO_HIA input XML file containing the prevalence of a disease
}
