% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeExcessMort}
\alias{writeExcessMort}
\title{######################   function to write     ###############################
#####################  disease Excess mortality   #############################
Writes the DYNAMO_HIA input XML file containing the excess mortality of a disease}
\usage{
writeExcessMort(
  diseasename,
  dynamodir,
  filename,
  data,
  type = "Rate",
  varname,
  acutely_fatal = NULL,
  cured_fraction = NULL
)
}
\arguments{
\item{diseasename}{Disease name as used in DYNAMO-HIA}

\item{dynamodir}{DYNAMO-HIA work directory}

\item{filename}{filename of the xml file containing the excess mortality data (without the extension .xml)}

\item{data}{dataframe with variables age, sex, and values to be used (with variable names as given by next parameters)
were sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{type}{one of "Rate" or "Median Survival"}

\item{varname}{name of the variable in the dataframe that contains the excess mortality values to be put
in the file, namely rate or median survival}

\item{acutely_fatal}{name of the variable in the dataframe that contains the values for the acutely
fatal fraction (should be NULL (default when no acutely fatal fraction is present))
in the file, namely the DALY weight (values between 0 and 100)}

\item{cured_fraction}{name of the variable in the dataframe that contains the values for the acutely
fatal fraction (values between 0 and 100). (should be NULL (default when no cured fraction is present))}
}
\description{
######################   function to write     ###############################
#####################  disease Excess mortality   #############################
Writes the DYNAMO_HIA input XML file containing the excess mortality of a disease
}
