% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeNetTransition}
\alias{writeNetTransition}
\title{####################  function to write transition files           #####################
####################  nett for cat and duration                      #####################}
\usage{
writeNetTransition(RFname, dynamodir, filename)
}
\arguments{
\item{RFname}{riskfactorname as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{dynamodir}{DYNAMO-HIA work directory}

\item{filename}{filename of the xml file to write the prevalence to (without the extension .xml)}
}
\description{
Write the net transition file for a categorical or compound riskfactor
}
