% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeNetTransitionCont}
\alias{writeNetTransitionCont}
\title{write the nett-transitions file (nett drift) for a continuous riskfactor}
\usage{
writeNetTransitionCont(RFname, dynamodir, filename, trend)
}
\arguments{
\item{RFname}{riskfactorname as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{dynamodir}{DYNAMO-HIA work directory (full path)}

\item{filename}{filename of the xml file to write the transition data (without the extension .xml)}

\item{trend}{trend to be applied on top of a netto drift}
}
\description{
write the nett-transitions file (nett drift) for a continuous riskfactor
}
