% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeNewborn}
\alias{writeNewborn}
\title{######################   function to write     ###############################
#######################  newborns ######################################
writes the DYNAMO-HIA input file containing the number of newborns}
\usage{
writeNewborn(dynamodir, popdir, sexratio, startingyear, data)
}
\arguments{
\item{dynamodir}{DYNAMO-HIA work directory}

\item{popdir}{name of population}

\item{sexratio}{ratio of male versus female children born}

\item{startingyear}{the first number in the file give the newborns for this year}

\item{data}{a vector (one dimensional array) containing the newborn data starting in the year "startingyear"
only the first 30 elements are used. If it is shorter than 30, the other elements up to
30 are filled with the last value of the vector}
}
\description{
######################   function to write     ###############################
#######################  newborns ######################################
writes the DYNAMO-HIA input file containing the number of newborns
}
