% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeORdisability}
\alias{writeORdisability}
\title{Writes the DYNAMO_HIA input XML file containing the odds ratios (ORs) of a risk factor on disability}
\usage{
writeORdisability(
  RFname,
  ncat,
  dynamodir,
  filename,
  data,
  varname,
  compound = FALSE
)
}
\arguments{
\item{RFname}{= name of the risk factor as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{ncat}{number of categories in the riskfactor}

\item{dynamodir}{= DYNAMO-HIA work directory}

\item{filename}{filename under which to write the OR on disability XML file (without the riskfactor name and XML extension)}

\item{data}{dataframe with variables cat (categories of riskfactor), age, sex, value to be used (as given by varname)
where cat is coded 1 to ncat (consecutive) sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{name of variable that contains the value to put in file, namely the odds ratios from the risk factor on disability}
}
\description{
Writes the DYNAMO_HIA input XML file containing the odds ratios (ORs) of a risk factor on disability
}
\examples{
to do
## Write the file with odds ratios for (all cause) disability for the 3 category risk factor RF to the file
testfile-RF.xml at the right place in the dynamo-hia home  directory C:\temp\test 
data3[1,1,]<-rep(1,96) # Odds ratio for category 1 in men
data3[1,2,]<-rep(1,96) # Odds ratio for category 1 in women
data3[2,1,]<-rep(1.1,96)  # Odds ratio for category 2 in men
data3[2,2,]<-rep(1.2,96)  # Odds ratio for category 2 in women
data3[3,1,]<-rep(1.3,96)  # Odds ratio for category 3 in men
data3[3,2,]<-rep(1.4,96)  # Odds ratio for category 3 in women
writeORdisability("RF",3,"c:\\\\temp/test","testfile",data3) 
 
}
