% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeORdisabilityCont}
\alias{writeORdisabilityCont}
\title{Writes the DYNAMO_HIA input XML file containing the odds ratios (ORs) of a risk factor on disability}
\usage{
writeORdisabilityCont(RFname, dynamodir, filename, data, varname)
}
\arguments{
\item{RFname}{= name of the risk factor as used in DYNAMO-HIA (spaces will be replace by underscore)}

\item{dynamodir}{= DYNAMO-HIA work directory}

\item{filename}{filename under which to write the OR on disability XML file (without the riskfactor name and XML extension)}

\item{data}{dataframe with variables cat (categories of riskfactor), age, sex, value (of the oddsratio) to be used (as given by varname)
where sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{name of variable that contains the value to put in file, namely the odds ratios from the risk factor on disability}
}
\description{
Writes the DYNAMO_HIA input XML file containing the odds ratios (ORs) of a risk factor on disability
}
\examples{
to do

 
}
