% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeOverallDisability}
\alias{writeOverallDisability}
\title{######################   function to write     ###############################
#######################  total disability    #################################
Function to write the DYNAMO_HIA input XML file for Overall Disability}
\usage{
writeOverallDisability(dynamodir, popdir, data, varname)
}
\arguments{
\item{dynamodir}{= DYNAMO-HIA work directory}

\item{popdir}{= name of population}

\item{data}{dataframe with variables age, sex, value to be used (as given by varname)
were sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{namely the overall disability in the population
in the file}
}
\description{
######################   function to write     ###############################
#######################  total disability    #################################
Function to write the DYNAMO_HIA input XML file for Overall Disability
}
