% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeOverallMortality}
\alias{writeOverallMortality}
\title{######################   function to write     ###############################
######################    total mortality      ###############################
writes the DYNAMO_HIA input XML file containing the rates of all cause mortality}
\usage{
writeOverallMortality(dynamodir, popdir, data, varname)
}
\arguments{
\item{dynamodir}{= DYNAMO-HIA work directory}

\item{popdir}{= name of population}

\item{data}{dataframe with variables age, sex, value to be used (as given by varname)
were sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{namely the all cause mortality rate
in the file}
}
\description{
######################   function to write     ###############################
######################    total mortality      ###############################
writes the DYNAMO_HIA input XML file containing the rates of all cause mortality
}
