% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeRiskFactorsDurationPrev}
\alias{writeRiskFactorsDurationPrev}
\title{function to write prevalence filesfor the duration category of a compound risk factor

Writes the DYNAMO_HIA input XML file  containing the riskfactor prevalence of a categorical (or compound) risk factor}
\usage{
writeRiskFactorsDurationPrev(RFname, dynamodir, data, varname)
}
\arguments{
\item{RFname}{= Risk factor name as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{dynamodir}{= DYNAMO-HIA work directory}

\item{data}{dataframe with variables age, sex, duration and value to be used (as given by varname)
where sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+ and
duration has values 1 to 20, where 20 is 20 years and longer}

\item{varname, }{list with the name of variables (one for each risk factor category) that contains
percentage in the duration class (values between 0-100)}
}
\value{
the function writes a file and does not return
}
\description{
function to write prevalence filesfor the duration category of a compound risk factor

Writes the DYNAMO_HIA input XML file  containing the riskfactor prevalence of a categorical (or compound) risk factor
}
