% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeRiskFactorsPrevCategorical}
\alias{writeRiskFactorsPrevCategorical}
\title{#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
Writes the DYNAMO_HIA input XML file  containing the riskfactor prevalence of a categorical (or compound) risk factor}
\usage{
writeRiskFactorsPrevCategorical(
  RFname,
  ncat,
  dynamodir,
  filename,
  data,
  varnames,
  referenceclass = 1,
  compound = FALSE,
  durationclass = NULL
)
}
\arguments{
\item{RFname}{= Risk factor name as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{ncat}{= number of categories in the riskfactor}

\item{dynamodir}{= DYNAMO-HIA work directory}

\item{filename}{= name to give to the riskfactor prevalence XML file (without XML extension)}

\item{data}{dataframe with variables age, sex, value to be used (as given by varname)
where sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varnames, }{list with the names of variables (one for each risk factor category) that contains the value to put in file, namely the
risk factor prevalence in percent (values between 0-100)}

\item{referenceclass}{the number of the referenceclass (numbered in order as given in varnames).
If not given, the first category given in varnames is taken as the reference class}

\item{compound}{Boolean indicating whether this is a compound risk factor}

\item{durationclass}{the number of the category with a time dependent relative risk}
}
\value{
the function writes a file and does not return
}
\description{
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
Writes the DYNAMO_HIA input XML file  containing the riskfactor prevalence of a categorical (or compound) risk factor
}
\examples{
# writing the prevalence (5\%,10\%,85\% in men, 4\%,6\%,90\% in women) of a 3 category riskfactor
# as well as the risk factor configurationfile
# this is written to the file testfile.xml in the dynamo-hia home directory c:\temp\test
data3<-array(0,c(3,2,96))
data3[1,1,]<-rep(5.1,96) # prevalence category 1 in men
data3[1,2,]<-rep(4.1,96) # prevalence category 1 in women
data3[2,1,]<-rep(10,96)  # prevalence category 2 in men
data3[2,2,]<-rep(6,96)   # prevalence category 2 in women
data3[3,1,]<-rep(85,96)  # prevalence category 3 in men
data3[3,2,]<-rep(90,96)  # prevalence category 3 in women
writeRiskFactorsPrev("RF",3,"c:/temp/test","testfile",data3) 
writeRiskFactorsPrev("RF",3,"/rivm/n/boshuizh/dynamo-hia/GUI","testfile",data3) 
}
