% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeRiskFactorsPrevCont}
\alias{writeRiskFactorsPrevCont}
\title{Writes the DYNAMO_HIA input XML file containing the prevalence of a continuous risk factor}
\usage{
writeRiskFactorsPrevCont(
  RFname,
  dynamodir,
  filename,
  data,
  varname,
  stdev,
  skewness = NULL,
  type = "Normal",
  referencevalue = NULL,
  cutoffs = NULL
)
}
\arguments{
\item{RFname}{riskfactorname as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{dynamodir}{DYNAMO-HIA work directory}

\item{filename}{filename of the xml file to write the prevalence to (without the extension .xml)}

\item{data}{dataframe with variables age, sex, and variable names as given below by varnames
#' were sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname}{name of the variable in the dataframe that contains the mean value of the}

\item{type}{one of "Rate" or "Median Survival"}

\item{cutoffs}{list with cutoff values to be used. Default: the mean over all ages}

\item{acutely_fatal}{name of the variable in the dataframe that contains the values for the acutely
fatal fraction (default: zero)
in the file, namely the DALY weight (values between 0 and 100)}

\item{cured_fraction}{name of the variable in the dataframe that contains the values for the acutely
fatal fraction (values between 0 and 100). Defaults to zero}

\item{isnormal}{boolean indicating whether the distribution is normal}
}
\description{
Writes the DYNAMO_HIA input XML file containing the prevalence of a continuous risk factor
}
