% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeRRdisease}
\alias{writeRRdisease}
\title{Writes the DYNAMO-HIA input xml file containing relative risks of a disease on another disease}
\usage{
writeRRdisease(
  diseasename,
  causaldiseasename,
  dynamodir,
  filename,
  data,
  varname
)
}
\arguments{
\item{diseasename}{Disease name as used in DYNAMO-HIA of the disease that is influenced by the causal disease}

\item{causaldiseasename}{Disease name as used in DYNAMO-HIA of the disease influencing the other disease}

\item{dynamodir}{= DYNAMO-HIA work directory (full path)}

\item{filename}{name to give to the relative risk XML file (without XML extension)}

\item{data}{dataframe with variables age, sex, value to be used (as given by varname)
were sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{namely the relative risk of the causal disease on the disease
in the file}
}
\description{
Writes the DYNAMO-HIA input xml file containing relative risks of a disease on another disease
}
