% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeRRMortalityCompound}
\alias{writeRRMortalityCompound}
\title{#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
####################  function to write  RR for a           #####################
####################  compound riskfactor on mortality      #####################
Writes the DYNAMO_HIA input XML file containing the relative risks (RR) from a riskfactor on all cause mortality}
\usage{
writeRRMortalityCompound(
  RFname,
  ncat,
  durcat,
  dynamodir,
  filename,
  data,
  varname,
  RRend,
  alpha
)
}
\arguments{
\item{RFname}{name if the risk factor as used in DYNAMO-HIA (any spaces will be replace by underscores)}

\item{ncat}{number of categories in the riskfactor}

\item{durcat}{number of the categories where the relative risk is time dependent}

\item{dynamodir}{DYNAMO-HIA work directory}

\item{filename}{name to give to the relative risk XML file (without the prescibed riskfactor name and the XML extension)}

\item{data}{dataframe with variables cat (categories of riskfactor), age, sex, value to be used (as given by varname)
where cat is coded 1 to ncat (consecutive) sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname}{name of variable that contains the relative risk on all cause mortality for each category of the risk factor
For the category where the relative risk is time dependent, this contains the relative risk at the start of exposure
that is, the moment the subject enters in this risk category}

\item{RRend}{name of the variable containing the relative risk after an infinitely long time in the risk category with}

\item{alpha}{name of the variable containing the alpha parameter (rate of change) of the time dependent relative risk
the time dependent relative risk}
}
\description{
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
####################  function to write  RR for a           #####################
####################  compound riskfactor on mortality      #####################
Writes the DYNAMO_HIA input XML file containing the relative risks (RR) from a riskfactor on all cause mortality
}
