% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeRRMortalityCont}
\alias{writeRRMortalityCont}
\title{Writes the DYNAMO_HIA input XML file containing the relative risks (RR) from a riskfactor on all cause mortality}
\usage{
writeRRMortalityCont(RFname, ncat, dynamodir, filename, data, varname)
}
\arguments{
\item{RFname}{= name if the risk factor as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{ncat}{number of categories in the riskfactor}

\item{dynamodir}{= DYNAMO-HIA work directory}

\item{filename}{= name to give to the relative risk XML file (without the prescibed riskfactor name and the XML extension)}

\item{data}{dataframe with variables cat (categories of riskfactor), age, sex, value to be used (as given by varname)
where cat is coded 1 to ncat (consecutive) sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{name of variable that contains the value to put in file, namely the
relative risk on all cause mortality for each category of the risk factor
#'}
}
\description{
Writes the DYNAMO_HIA input XML file containing the relative risks (RR) from a riskfactor on all cause mortality
}
\examples{
## writing the relative risks on mortality for a 3 category risk factor named RF
## this is written to the file RRmortFromStudyX-RF.xml in the right place in the 
## dynamo-hia home directory c:\temp\test
data3<-array(0,c(3,2,96))
data3[1,1,]<-rep(1,96) # relative risk for category 1 in men
data3[1,2,]<-rep(1,96) # relative risk for category 1 in women
data3[2,1,]<-rep(1.1,96)  # relative risk for category 2 in men
data3[2,2,]<-rep(1.15,96)   # relative risk for category 2 in women
data3[3,1,]<-rep(1.2,96)  # relative risk for category 3 in men
data3[3,2,]<-rep(1.3,96)  # relative risk for category 3 in women
writeRiskFactorsPrev("RF",3,"c:/temp/test","testfile",data3) 
writeRRMortality("RF",3,"c:/temp/test", "RRmortFromStudyX",  data2) 

}
