% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeRRriskfactorCat}
\alias{writeRRriskfactorCat}
\title{Writes the DYNAMO_HIA input XML file containing the relative risk of a categorical risk factor on diseae}
\usage{
writeRRriskfactorCat(
  diseasename,
  riskfactorname,
  ncat,
  dynamodir,
  filename,
  data,
  varname
)
}
\arguments{
\item{diseasename}{Disease name as used in DYNAMO-HIA}

\item{riskfactorname}{Risk Factor name as used in DYNAMO-HIA}

\item{ncat}{number of categories in the riskfactor}

\item{dynamodir}{= DYNAMO-HIA work directory}

\item{filename}{= name to give to the relative risk XML file (without XML extension and excluding
the risk factor name, this will be added automatically)}

\item{data}{dataframe with variables cat (categories of riskfactor), age, sex, value to be used (as given by varname)
where cat is coded 1 to ncat (consecutive) sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname, }{name of variable that contains the value to put in file, namely the
relative risk on the disease for each category of the risk factor}
}
\description{
Writes the DYNAMO_HIA input XML file containing the relative risk of a categorical risk factor on diseae
}
\examples{
todo
}
