% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeSimulationConfiguration}
\alias{writeSimulationConfiguration}
\title{####################################################################################write the simulation configuration file

#####################################################################################
Write the configuration file for simulation with simulationName}
\usage{
writeSimulationConfiguration(
  dynamodir,
  simulationName,
  newborns,
  startingyear,
  yearsSimulated,
  popSize,
  minAge,
  maxAge,
  refScenarioName,
  randomSeed,
  populationName,
  scenarioList,
  diseaseList,
  riskfactorList,
  relriskList
)
}
\arguments{
\item{dynamodir}{Dynamo-HIA workdirectory}

\item{simulationName}{Name of the simulation}

\item{newborns}{Boolean: False= newborns added during simulation True: newborns added}

\item{startingyear}{startyear of simulation}

\item{yearsSimulated}{number of years to simulate}

\item{popSize}{number of simulated persons per age and sex group}

\item{minAge}{minimum age of simualted population}

\item{maxAge}{maximum age of simulated population}

\item{refScenarioName}{name of the reference (business as usual) scenario}

\item{randomSeed}{random seed used in the simulation}

\item{populationName}{name of the population to be used}

\item{scenarioList}{list containing the information per scenario (not including the reference scenario)
each element contains information on a single scenario, with named elements:
name (name of scenario),
successRate: percentage of population to which the scenario is applied.
minAge,maxAge: minimum and maximum age for which the scenario is applied.
targetSex: should scenario be applied to men (0), women (1) or both (2)
transitionFile (name of the transition file for the risk factor,
prevalenceFile: name of the prevalence file for the risk factor

name, transitionfilename and prevalencefilename}

\item{diseaseList}{list with one element per disease. Each element has the following named elements:
name,
prevalenceFile: file with disease prevalence data,
incidenceFile:file with disease incidence data,
excessMortFile: file with diseaes excess mortality data,
disabilityFile: file with disease disability data}

\item{riskfactorList}{list with named elements:
name: name of the riskfactor,
transitionFile: name of the file with the transition for the riskfactor in the reference scenario,
prevalenceFile: name of the file with the starting prevalence for the riskfactor in the reference scenario}

\item{relriskList}{list with one element per relative risk. Each element has the following named elements:
name: name of the relative risk,
from: name of the causal exposure (risk factor or a disease causing another disease),
to: name of the effect (a disease, death or disability),
filename: name of the file containing the relative risks}
}
\description{
####################################################################################write the simulation configuration file

#####################################################################################
Write the configuration file for simulation with simulationName
}
