% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeSize}
\alias{writeSize}
\title{#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
###################### function to write ###############################
####################### population size ######################################
writes the DYNAMO-HIA input file containing the number of newborns}
\usage{
writeSize(dynamodir, popdir, data, varname)
}
\arguments{
\item{dynamodir}{DYNAMO-HIA work directory}

\item{popdir}{name of population}

\item{data}{a dataframe containing ages 0-95 (95 stands for 95+) and ses and a variable
given by varname containing the numbers in the population}

\item{varname}{name of the variable containing the numbers in the population}
}
\description{
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
#################################################################################
###################### function to write ###############################
####################### population size ######################################
writes the DYNAMO-HIA input file containing the number of newborns
}
