% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeTransitionMatrix}
\alias{writeTransitionMatrix}
\title{Writes the DYNAMO_HIA input XML file containing the transition probabilities for
a categorical or compound riskfactor}
\usage{
writeTransitionMatrix(RFname, dynamodir, filename, ncat, data, varname)
}
\arguments{
\item{RFname}{riskfactorname as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{dynamodir}{DYNAMO-HIA work directory}

\item{filename}{filename of the xml file to write the prevalence to (without the extension .xml)}

\item{ncat}{number of risk factor classes}

\item{data}{dataframe with variables age, sex, to and from and variable name as given below by varname
were sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+
from gives the category from which the transition goes to the category indicated by to
when to=from, the value of varname should be 100}

\item{varname}{name of the variable in the data.frame that contains the transition probability}
}
\description{
Writes the DYNAMO_HIA input XML file containing the transition probabilities for
a categorical or compound riskfactor
}
