% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeTransitionsCont}
\alias{writeTransitionsCont}
\title{Write the transitions (drift) for a continuous riskfactor}
\usage{
writeTransitionsCont(RFname, dynamodir, filename, data, varname)
}
\arguments{
\item{RFname}{riskfactorname as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{dynamodir}{DYNAMO-HIA work directory}

\item{filename}{filename of the xml file to write the prevalence to (without the extension .xml)}

\item{data}{dataframe with variables age, sex, and variable names as given below by varnames
#' were sex is coded as (0 = men, 1 = women) and age has values 0 - 95 where 95 stands for 95+}

\item{varname}{name in the dataset given above that contains the drift of the riskfactor
defined as the increase when the subject ages one year}
}
\value{
NULL a file is written
}
\description{
Write the transitions (drift) for a continuous riskfactor
}
