% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeZeroTransition}
\alias{writeZeroTransition}
\title{Write the zero transitions for a categorical or compound riskfactor}
\usage{
writeZeroTransition(RFname, dynamodir, filename)
}
\arguments{
\item{RFname}{riskfactorname as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{dynamodir}{DYNAMO-HIA work directory}

\item{filename}{filename of the xml file to write the transition rates (without the extension .xml)}
}
\description{
Write the zero transitions for a categorical or compound riskfactor
}
