% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefunctions_for_DYNAMO.R
\name{writeZeroTransitionCont}
\alias{writeZeroTransitionCont}
\title{Write the zero transitions (drift) for a continuous riskfactor}
\usage{
writeZeroTransitionCont(RFname, dynamodir, filename)
}
\arguments{
\item{RFname}{riskfactorname as used in DYNAMO-HIA. Any spaces in the name will be replaced by _}

\item{dynamodir}{DYNAMO-HIA work directory}

\item{filename}{filename of the xml file to write the transition rates (without the extension .xml)}
}
\description{
Write the zero transitions (drift) for a continuous riskfactor
}
