% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading_helpers.R
\name{configure_scenario}
\alias{configure_scenario}
\title{Configure a scenario}
\usage{
configure_scenario(
  name,
  success_rate,
  min_age,
  max_age,
  gender,
  transition_filename,
  prevalence_filename
)
}
\arguments{
\item{name}{A character string which specifies the unique name of the scenario.}

\item{success_rate}{A numeric which specifies the intervention success rate (0-100).}

\item{min_age}{A numeric which specifies the minimum target age (0-95).}

\item{max_age}{A numeric which specifies the maximum target age (0-95).}

\item{gender}{A numeric which specifies the target gender (0=both, 1=male, 2=female).}

\item{transition_filename}{A character string which specifies the path to the
transition rates file.}

\item{prevalence_filename}{A character string which specifies the path to the prevalence file.}
}
\value{
A list containing the scenario configuration parameters.
}
\description{
Creates a configuration list for an intervention scenario in the simulation.
}
\examples{
configure_scenario(
  name = "All_Never_Smokers",
  success_rate = 80,
  min_age = 18,
  max_age = 65,
  gender = 2, # Both female and male
  transition_filename = "NL_RF_smoking_Transitions_Netto",
  prevalence_filename = "RF_smoking_cat_All_Never_Smokers_Prev"
)

}
