% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_module.R
\name{single_scenario_server}
\alias{single_scenario_server}
\title{Server logic for defining a single scenario}
\usage{
single_scenario_server(id, reference_data)
}
\arguments{
\item{id}{A character string. The ID used to namespace the server-side logic.
It should match the id of the corresponding UI elements.}

\item{reference_data}{A reactive expression. The expression determines the return
value of the server.}
}
\value{
If \code{reference_data} returns \code{NULL}, the server returns an empty list.
Otherwise, it returns a list where each element is taken from the UI inputs.
}
\description{
Server logic for defining a single scenario
}
\keyword{internal}
