% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_module.R
\name{single_scenario_ui}
\alias{single_scenario_ui}
\title{Create UI for single scenario}
\usage{
single_scenario_ui(id, prevalence_and_transition_choices)
}
\arguments{
\item{id}{A character string. The ID used to namespace the UI elements.}

\item{prevalence_and_transition_choices}{A reactive value that returns a list
of two character vector of choices. The names should be \code{prevalences} and
\code{transitions}.}
}
\value{
A Shiny UI component for defining a scenario.
}
\description{
Create UI for single scenario
}
\keyword{internal}
